Refactored "onJumpscareAttempt" to "onTick".
I think this is more clear, since this is always called on each Animatronic on the map each tick. Also, I opened the door for new information to be recalculated each tick and communicated to the Night on future Animatronics. Probably useless, but I did it to make it more clear that the return of the onTick is information about whether to start the jumpscare state on that tick for that Animatronic or not.