With absolute chaos ensured, Jumpscares now truly read the GIF's information on delays between each image.
That means it will be easier for me to create the animations. Nonetheless, 100% recommended that each frame is correctly sized, and also when the window is taller than wider it displays centered. I need to fix that somehow.

For the curious, prior to this the game would only read each frame of the animation and place it on the screen based on a number of repetitions, which also did not depend on FPS leading to faster animations on higher FPS and slower animations on lower FPS. Now it should all be fine!