Published Maven Package with version 0.7.8 of the project.
Kind of for testing, but perhaps I can also try doing "modifications" in a sepparate project using this one as a dependency. Maybe I could do it for the most wacky stuff, like future custom AnimatronicDrawings for Custom Night that may not make the cut for the main release while designing them.