NightJC now asks for BufferedImage and Sound instead of their paths to resources.
Again, this is to allow assets to be loaded from somewhere else if desired. It is important to note that assets used by the Nights are loaded from resources and will not be modificable by using it as a dependency since it would need a lot more changes (like a Builder or something).